+++ CHANGELOG.md
@@ -11,3 +11,4 @@ All notable changes to this project will be documented in this file.
 ## 0.1.1
 ### Added
 - Simple example `helloworld` project under `examples/` ([#62](https://github.com/google/jib/pull/62))
+- Infers common credential helper names (for GCR and ECR) ([#64](https://github.com/google/jib/pull/64))